+++ spec/kuber_kit/service_deployer/strategies/kubernetes_spec.rb
@@ -19,16 +19,31 @@ RSpec.describe KuberKit::ServiceDeployer::Strategies::Kubernetes do
     expect(subject.kubectl_commands).to receive(:rolling_restart).with(
       shell, "deployment", "auth-app", kubeconfig_path: nil, namespace: nil
     )
+    expect(subject.kubectl_commands).to receive(:rollout_status).with(
+      shell, "deployment", "auth-app", wait: true, kubeconfig_path: nil, namespace: nil
+    )
 
     subject.deploy(shell, service)
   end
 
   it "doesn't restart deployment if it's disabled for service" do
     expect(shell).to receive(:write).with("/tmp/kuber_kit/services/auth_app.yml", /apiVersion: v1/)
+    expect(subject.kubectl_commands).to receive(:resource_exists?).and_return(true)
     expect(subject.kubectl_commands).to receive(:apply_file)
     expect(subject.kubectl_commands).to_not receive(:rolling_restart)
 
     service = service_helper.service(:auth_app, attributes: {deployer: {restart_if_exists: false}})
+    subject.deploy(shell, service)
+  end
+
+  it "doesn't wait for rollout status deployment if it's disabled for service" do
+    expect(shell).to receive(:write).with("/tmp/kuber_kit/services/auth_app.yml", /apiVersion: v1/)
+    expect(subject.kubectl_commands).to receive(:resource_exists?).and_return(true)
+    expect(subject.kubectl_commands).to receive(:apply_file)
+    expect(subject.kubectl_commands).to receive(:rolling_restart)
+    expect(subject.kubectl_commands).to_not receive(:rollout_status)
+
+    service = service_helper.service(:auth_app, attributes: {deployer: {restart_if_exists: true, wait_for_rollout: false}})
     subject.deploy(shell, service)
   end