+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 **0.5.2**
 - Added dependencies support to services
 - Added an option to deploy all services in `kit deloy`
+- Wait for rollout to finish during service deploy